vsphere-vm-resource: refactor out functions for shared use
This refactors out some functions needed for implementing the vSphere
K8s cluster provider so they can be shared between the two crates
