+++ bottlerocket/agents/src/bin/vsphere-vm-resource-agent/main.rs
@@ -5,7 +5,6 @@ Provides Bottlerocket VMWare vSphere VMs to serve as Kubernetes nodes via `govc`
 !*/
 
 mod aws;
-mod tuf;
 mod vsphere_vm_provider;
 
 use crate::vsphere_vm_provider::{VMCreator, VMDestroyer};
